+++ script.module.libmediathek3/addon.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
 <addon id="script.module.libmediathek3" 
 	     name="libmediathek3" 
-	     version="1.3.0" 
+	     version="1.3.1" 
 	     provider-name="membrane">
   <requires>
     <import addon="xbmc.python" version="2.25.0" />
