+++ script.module.libmediathek3/changelog.txt
@@ -1,3 +1,5 @@
+1.3.1
+- Fixed problem with predefined directories
 1.3.0
 - Added workaround for Krypton caching quirks
 1.2.9
