+++ script.module.libmediathek3/lib/libmediathek3dialogs.py
@@ -12,8 +12,17 @@ import re
 from datetime import datetime,timedelta
 import time
 
-from libmediathek3utils import getTranslation as translation
+import libmediathek3utils
 
 def dialogDate():
 	dialog = xbmcgui.Dialog()
-	return dialog.numeric(1, translation(31030)).replace('/','').replace(' ','0')
+	return dialog.numeric(1, libmediathek3utils.getTranslation(31030)).replace('/','').replace(' ','0')
+	
+def getSearchString():
+	if libmediathek3utils.searchWorkaroundExists():
+		d = libmediathek3utils.searchWorkaroundRead()
+	else:
+		dialog = xbmcgui.Dialog()
+		d = dialog.input(libmediathek3utils.getTranslation(31039),type=xbmcgui.INPUT_ALPHANUM)
+		libmediathek3utils.searchWorkaroundWrite(d)
+	return urllib.quote_plus(d)
